+++ testnet/stacks-node/src/tests/epoch_23.rs
@@ -18,10 +18,10 @@ use std::{env, thread};
 
 use clarity::vm::types::{PrincipalData, QualifiedContractIdentifier};
 use stacks::burnchains::{Burnchain, PoxConstants};
+use stacks::config::InitialBalance;
 use stacks::core::{self, EpochList, STACKS_EPOCH_MAX};
 use stacks_common::util::sleep_ms;
 
-use crate::config::InitialBalance;
 use crate::tests::bitcoin_regtest::BitcoinCoreController;
 use crate::tests::neon_integrations::*;
 use crate::tests::*;